[REMANIEMENT][RESTITUTION] Sépare la génération de la restitution pour les versions HTML et PDF
Le rendu n'est plus le même chaque implémentation doit gérer sont propre rendu
