+++ mon-aide-cyber-api/src/adaptateurs/AdaptateurDeRestitution.ts
@@ -4,7 +4,6 @@ import {
   ORDRE_THEMATIQUES,
 } from '../diagnostic/Diagnostic';
 import { Restitution } from '../restitution/Restitution';
-import { FournisseurHorloge } from '../infrastructure/horloge/FournisseurHorloge';
 
 export const estMesurePrioritaire = (
   mesurePrioritaire: MesurePriorisee[] | undefined
@@ -15,10 +14,11 @@ export const estMesurePrioritaire = (
 export type ContenuHtml = { corps: string; entete: string; piedPage: string };
 
 export abstract class AdaptateurDeRestitution<T> {
-  constructor(private readonly traductionThematiques: Map<string, string>) {}
+  constructor(protected readonly traductionThematiques: Map<string, string>) {}
 
-  genereRestitution(restitution: Restitution): Promise<T> {
-    const informations = this.genereInformations(restitution);
+  public abstract genereRestitution(restitution: Restitution): Promise<T>;
+
+  protected trieLesIndicateurs(restitution: Restitution) {
     const indicateursRestitution: Indicateurs = Object.entries(
       restitution.indicateurs
     )
@@ -35,108 +35,6 @@ export abstract class AdaptateurDeRestitution<T> {
         }),
         {}
       );
-    const indicateurs = this.genereIndicateurs(indicateursRestitution);
-    const mesuresPrioritaires = this.genereMesuresPrioritaires(
-      restitution.mesures.mesuresPrioritaires
-    );
-    const autresMesures = restitution.mesures.autresMesures;
-
-    const contactsLiensUtiles = this.genereContactsEtLiensUtiles();
-    const ressources = this.genereRessources();
-
-    return this.genereToutesLesPages(
-      autresMesures,
-      informations,
-      indicateurs,
-      mesuresPrioritaires,
-      contactsLiensUtiles,
-      ressources
-    );
-  }
-
-  protected genereToutesLesPages(
-    autresMesures: MesurePriorisee[],
-    informations: Promise<ContenuHtml>,
-    indicateurs: Promise<ContenuHtml>,
-    mesuresPrioritaires: Promise<ContenuHtml>,
-    contactsLiensUtiles: Promise<ContenuHtml>,
-    ressources: Promise<ContenuHtml>
-  ) {
-    if (estMesurePrioritaire(autresMesures)) {
-      return this.genere([
-        informations,
-        indicateurs,
-        mesuresPrioritaires,
-        contactsLiensUtiles,
-        ressources,
-        this.genereAutresMesures(autresMesures),
-      ]);
-    }
-    return this.genere([
-      informations,
-      indicateurs,
-      mesuresPrioritaires,
-      contactsLiensUtiles,
-      ressources,
-    ]);
-  }
-
-  protected abstract genere(mesures: Promise<ContenuHtml>[]): Promise<T>;
-
-  protected genereIndicateurs(
-    indicateurs: Indicateurs | undefined
-  ): Promise<ContenuHtml> {
-    return this.genereHtml('indicateurs', {
-      indicateurs,
-      traductions: this.traductionThematiques,
-    });
-  }
-
-  abstract genereHtml(pugCorps: string, paramsCorps: any): Promise<ContenuHtml>;
-
-  protected async genereInformations(
-    restitution: Restitution
-  ): Promise<ContenuHtml> {
-    return this.genereHtml('informations', {
-      ...this.representeInformations(restitution),
-    });
-  }
-
-  private representeInformations(restitution: Restitution) {
-    return {
-      dateCreation: FournisseurHorloge.formateDate(
-        restitution.informations.dateCreation
-      ),
-      dateDerniereModification: FournisseurHorloge.formateDate(
-        restitution.informations.dateDerniereModification
-      ),
-      identifiant: restitution.identifiant,
-      secteurActivite: restitution.informations.secteurActivite,
-      secteurGeographique: restitution.informations.secteurGeographique,
-    };
-  }
-
-  protected genereAutresMesures(
-    autresMesures: MesurePriorisee[] | undefined
-  ): Promise<ContenuHtml> {
-    return this.genereHtml('autres-mesures', {
-      mesures: autresMesures,
-    });
-  }
-
-  protected genereMesuresPrioritaires(
-    mesuresPrioritaires: MesurePriorisee[] | undefined
-  ): Promise<ContenuHtml> {
-    return this.genereHtml('mesures', {
-      mesures: mesuresPrioritaires,
-    });
-  }
-
-  protected genereContactsEtLiensUtiles(): Promise<ContenuHtml> {
-    return this.genereHtml('contacts-liens-utiles', {});
-  }
-
-  protected genereRessources(): Promise<ContenuHtml> {
-    return this.genereHtml('ressources', {});
+    return indicateursRestitution;
   }
 }
