+++ mon-aide-cyber-api/src/infrastructure/adaptateurs/AdaptateurDeRestitutionHTML.ts
@@ -2,7 +2,11 @@ import * as pug from 'pug';
 import {
   AdaptateurDeRestitution,
   ContenuHtml,
+  estMesurePrioritaire,
 } from '../../adaptateurs/AdaptateurDeRestitution';
+import { Restitution } from '../../restitution/Restitution';
+import { Indicateurs, MesurePriorisee } from '../../diagnostic/Diagnostic';
+import { FournisseurHorloge } from '../horloge/FournisseurHorloge';
 
 export type RestitutionHTML = {
   autresMesures: string;
@@ -16,6 +20,99 @@ export type RestitutionHTML = {
 export class AdaptateurDeRestitutionHTML extends AdaptateurDeRestitution<RestitutionHTML> {
   constructor(traductionThematiques: Map<string, string>) {
     super(traductionThematiques);
+  }
+
+  public genereRestitution(restitution: Restitution): Promise<RestitutionHTML> {
+    return this.genereLaRestitution(restitution);
+  }
+
+  protected genereLaRestitution(
+    restitution: Restitution
+  ): Promise<RestitutionHTML> {
+    const informations = this.genereInformations(restitution);
+    const indicateursRestitution: Indicateurs =
+      this.trieLesIndicateurs(restitution);
+    const indicateurs = this.genereIndicateurs(indicateursRestitution);
+    const mesuresPrioritaires = this.genereMesuresPrioritaires(
+      restitution.mesures.mesuresPrioritaires
+    );
+    const autresMesures = restitution.mesures.autresMesures;
+
+    const contactsLiensUtiles = this.genereContactsEtLiensUtiles();
+    const ressources = this.genereRessources();
+
+    if (estMesurePrioritaire(autresMesures)) {
+      return this.genere([
+        informations,
+        indicateurs,
+        mesuresPrioritaires,
+        contactsLiensUtiles,
+        ressources,
+        this.genereAutresMesures(autresMesures),
+      ]);
+    }
+    return this.genere([
+      informations,
+      indicateurs,
+      mesuresPrioritaires,
+      contactsLiensUtiles,
+      ressources,
+    ]);
+  }
+
+  protected genereIndicateurs(
+    indicateurs: Indicateurs | undefined
+  ): Promise<ContenuHtml> {
+    return this.genereHtml('indicateurs', {
+      indicateurs,
+      traductions: this.traductionThematiques,
+    });
+  }
+
+  protected async genereInformations(
+    restitution: Restitution
+  ): Promise<ContenuHtml> {
+    return this.genereHtml('informations', {
+      ...this.representeInformations(restitution),
+    });
+  }
+
+  protected representeInformations(restitution: Restitution) {
+    return {
+      dateCreation: FournisseurHorloge.formateDate(
+        restitution.informations.dateCreation
+      ),
+      dateDerniereModification: FournisseurHorloge.formateDate(
+        restitution.informations.dateDerniereModification
+      ),
+      identifiant: restitution.identifiant,
+      secteurActivite: restitution.informations.secteurActivite,
+      secteurGeographique: restitution.informations.secteurGeographique,
+    };
+  }
+
+  protected genereAutresMesures(
+    autresMesures: MesurePriorisee[] | undefined
+  ): Promise<ContenuHtml> {
+    return this.genereHtml('autres-mesures', {
+      mesures: autresMesures,
+    });
+  }
+
+  protected genereMesuresPrioritaires(
+    mesuresPrioritaires: MesurePriorisee[] | undefined
+  ): Promise<ContenuHtml> {
+    return this.genereHtml('mesures', {
+      mesures: mesuresPrioritaires,
+    });
+  }
+
+  protected genereContactsEtLiensUtiles(): Promise<ContenuHtml> {
+    return this.genereHtml('contacts-liens-utiles', {});
+  }
+
+  protected genereRessources(): Promise<ContenuHtml> {
+    return this.genereHtml('ressources', {});
   }
 
   protected async genere(
